View Component
Create View component & implement page titles
