+++ core/Application.php
@@ -16,6 +16,7 @@ class Application
     public Session $session;
     public Database $db;
     public ?DbModel $user;
+    public ?View $view;
 
     public static Application $app;
     public ?Controller $controller = null;
@@ -29,6 +30,7 @@ class Application
         $this->response = new Response();
         $this->session = new Session();
         $this->router = new Router($this->request, $this->request);
+        $this->view = new View();
 
         $this->db = new Database($config['db']);
 
@@ -52,7 +54,7 @@ class Application
             echo $this->router->resolve();
         } catch (\Exception $e) {
             $this->response->setStatusCode($e->getCode());
-            echo $this->router->renderView('_error', [
+            echo $this->view->renderView('_error', [
                 'exception' => $e
             ]);
         }
