+++ core/View.php
@@ -1,0 +1,39 @@
+<?php
+
+
+namespace app\core;
+
+
+class View
+{
+    public string $title = '';
+
+    public function renderView($view, $params = [])
+    {
+        $viewContent = $this->renderOnlyView($view, $params);
+        $layoutContent = $this->layoutContent();
+        return str_replace('{{content}}', $viewContent, $layoutContent);
+    }
+
+    private function renderContent($viewContent)
+    {
+        $layoutContent = $this->layoutContent();
+        return str_replace('{{content}}', $viewContent, $layoutContent);
+    }
+
+    protected function layoutContent() {
+        $layout = Application::$app->controller->layout;
+        ob_start();
+        include_once Application::$ROOT_DIR . "/views/layouts/$layout.php";
+        return ob_get_clean();
+    }
+
+    protected function renderOnlyView(string $view, array $params) {
+        foreach ($params as $key => $value) {
+            $$key = $value;
+        }
+        ob_start();
+        include_once Application::$ROOT_DIR . "/views/$view.php";
+        return ob_get_clean();
+    }
+}
