+++ views/profile.php
@@ -1,17 +1,4 @@
+<?php
+    $this->title = 'Profile'
+?>
 <h1>Profile</h1>
-<form class="container" method="post">
-    <h1>Contact Form</h1>
-    <div class="mb-3">
-        <label class="form-label">Subject</label>
-        <input type="text" class="form-control" name="subject">
-    </div>
-    <div class="mb-3">
-        <label class="form-label">Email</label>
-        <input type="text" class="form-control" name="email">
-    </div>
-    <div class="mb-3">
-        <label for="exampleInputEmail1" class="form-label">Body</label>
-        <textarea class="form-control" name="body"></textarea>
-    </div>
-    <button type="submit" class="btn btn-primary">Submit</button>
-</form>
